Sprint5_v1.7_editDeleteFile&Folder

+++ src/main/java/sit/int221/sas/sit_announcement_system_backend/config/JwtTokenUtil.java
@@ -128,7 +128,7 @@ public class JwtTokenUtil {
     }
 
     public  Boolean validateOtpEmail(Integer otp,String token,Claims claims){
-        return  ((Objects.equals((Integer) claims.get("otp"), otp)) && !isTokenExpired(token));
+        return  ((Objects.equals(claims.get("otp"), otp)) && !isTokenExpired(token));
     }
